refactor(gameMap): rename sidebar to reaarangeButton.
